restructure index
