+++ src/lib/component/fc-eval.ts
@@ -1,7 +1,94 @@
+import * as core from '@serverless-devs/core';
+import * as yaml from 'js-yaml';
+import * as HELP from '../help/eval';
 import { EvalOption } from '../interface/component/fc-eval';
 import { HttpTypeOption, PayloadOption } from '../interface/component/fc-common';
+import logger from '../../common/logger';
+import { isHttpFunction } from '../utils';
+import { IInputs } from '../interface/interface';
 
 export default class FcEval {
+  static async handlerComponentInputs(inputs: IInputs): Promise<any> {
+    const { props, project } = inputs;
+    const SUPPORTED_METHOD: string[] = ['start'];
+    const DEFAULT_EVAL_TYPE = 'memory';
+
+    const apts = {
+      boolean: ['help', 'assume-yes'],
+      alias: {
+        help: 'h',
+        region: 'r',
+        access: 'a',
+        'payload-file': 'f',
+        'assume-yes': 'y',
+      },
+    };
+    const args: string = (inputs?.args || '').replace(/(^\s*)|(\s*$)/g, '');
+    const comParse: any = core.commandParse({ ...inputs, args }, apts);
+    const argsData: any = comParse?.data || {};
+    const nonOptionsArgs = argsData?._ || [];
+
+    logger.debug(`nonOptionsArgs is ${JSON.stringify(nonOptionsArgs)}`);
+    if (!argsData) {
+      logger.error('Not fount sub-command.');
+      core.help(HELP.EVAL);
+      return { isHelp: true };
+    }
+    if (nonOptionsArgs.length === 0) {
+      if (!argsData?.help) {
+        logger.error('Not fount sub-command.');
+      }
+      core.help(HELP.EVAL);
+      return { isHelp: true };
+    }
+
+    const commandName: string = nonOptionsArgs[0];
+    if (!SUPPORTED_METHOD.includes(commandName)) {
+      logger.error(`Not supported sub-command: [${commandName}]`);
+      core.help(HELP.EVAL);
+      return { isHelp: true };
+    }
+
+    if (argsData?.help) {
+      core.help(HELP.EVAL_START);
+      return { isHelp: true };
+    }
+    let functionType = argsData['function-type'];
+    // yaml 模式
+    if (props?.service?.name && props?.function?.name) {
+      functionType = functionType || isHttpFunction(props) ? 'http' : 'event';
+    }
+
+    const evalOpts: EvalOption = {
+      serviceName: argsData['service-name'] || props?.service?.name,
+      functionName: argsData['function-name'] || props?.function?.name,
+      functionType,
+      evalType: argsData['eval-type'] || DEFAULT_EVAL_TYPE,
+      memorySizeList: argsData['memory-size'],
+      runCount: argsData['run-count'],
+      rt: argsData?.rt,
+      memory: argsData?.memory,
+      concurrencyArgs: argsData['concurrency-args'],
+    };
+
+    const httpTypeOpts: HttpTypeOption = {
+      url: argsData?.url,
+      method: argsData?.method,
+      path: argsData?.path,
+      query: argsData?.query,
+      body: argsData?.body,
+      headers: argsData?.headers,
+    };
+    logger.debug(`Using http options: \n${yaml.dump(httpTypeOpts)}`);
+    const payloadOpts: PayloadOption = {
+      payloadFile: argsData['payload-file'],
+      payload: argsData?.payload,
+    };
+
+    const region = props?.region || argsData?.region;
+    return { project, evalOpts, httpTypeOpts, payloadOpts, region, commandName }
+  }
+
   private readonly httpTypeOpts?: HttpTypeOption;
   private readonly evalOpts?: EvalOption;
   private readonly payloadOpts?: PayloadOption;
