+++ src/lib/interface/profile.ts
@@ -1,4 +1,5 @@
 export interface ICredentials {
+  Alias?: string;
   AccountID: string;
   AccessKeyID: string;
   AccessKeySecret: string;
